1. 去掉powershell可以直接执行！；2. 下载前检查安装包是否存在

+++ autoinit.py
@@ -76,13 +76,14 @@ def install_app(apps: list[dict, ...], download_dir: str = f'C:/Users/{os.getlog
         match app['method']:
             case 'automatic':
                 file_name = app['url'].split('/')[-1]
-                response = requests.get(app['url'], stream=True)
-                length = int(response.headers.get('content-length', 0))
-                with open(download_dir + file_name, 'wb') as fo, tqdm.tqdm(desc=file_name, total=length, unit='iB', unit_scale=True, unit_divisor=1024) as bar:
-                    for data in response.iter_content(chunk_size=1024):
-                        size = fo.write(data)
-                        bar.update(size)
-                os.system(f'PowerShell {download_dir + file_name} {app['args']}')
+                if not os.path.exists(download_dir + file_name):
+                    response = requests.get(app['url'], stream=True)
+                    length = int(response.headers.get('content-length', 0))
+                    with open(download_dir + file_name, 'wb') as fo, tqdm.tqdm(desc=file_name, total=length, unit='iB', unit_scale=True, unit_divisor=1024) as bar:
+                        for data in response.iter_content(chunk_size=1024):
+                            size = fo.write(data)
+                            bar.update(size)
+                os.system(f'{download_dir + file_name} {app['args']}')
                 input(COLOR_INFO + "Wait for the installation to complete.")
             case 'manual':
                 webbrowser.open(app['site'])
